requirements.txt
annotated-types==0.7.0
google-genai
anthropic==0.50.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.2.25
chardet==7.2.0
click==8.3.1
colorama==0.4.6
distro==1.9.0
Flask==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
MarkupSafe==3.0.3
pillow==12.1.1
psycopg2-binary==2.9.10
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.1.0
reportlab==4.4.1
sniffio==1.3.1
SQLAlchemy==2.0.48
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.6
WTForms==3.2.1
cloudinary==1.36.0
